fix: 페이먼츠 엔티티 수정

+++ module-api/src/main/java/com/whiskey/payment/controller/PaymentController.java
@@ -39,7 +39,7 @@ public class PaymentController {
         PaymentPrepareCommand command = new PaymentPrepareCommand(
             memberId,
             request.orderId(),
-            BigDecimal.valueOf(request.amount()),
+            request.amount(),
             request.description()
         );
