+++ module-domain/src/main/java/com/whiskey/domain/payment/dto/PaymentPrepareCommand.java
@@ -1,10 +1,8 @@
 package com.whiskey.domain.payment.dto;
 
-import java.math.BigDecimal;
-
 public record PaymentPrepareCommand(
     Long memberId,
     Long orderId,
-    BigDecimal amount,
+    Long amount,
     String description
 ) {}
