+++ module-domain/src/main/java/com/whiskey/domain/payment/dto/PaymentPrepareResult.java
@@ -1,8 +1,6 @@
 package com.whiskey.domain.payment.dto;
 
-import java.math.BigDecimal;
-
 public record PaymentPrepareResult(
     String orderId,
-    BigDecimal amount
+    Long amount
 ) {}
